# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ODIN Swagger UI",
        "title": "Odin: Remote Administration for 1S.Enterprise Application Servers",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "khorevaa@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/app/{app}/clusters/{cluster}/infobases": {
            "get": {
                "description": "получение списка информационных баз с кластера",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infobases"
                ],
                "summary": "получение списка информационных баз с кластера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serialize.InfobaseSummaryInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "description": "оздание информационной базы на кластере",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infobases"
                ],
                "summary": "оздание информационной базы на кластере",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "create server db",
                        "name": "create-db",
                        "in": "query"
                    },
                    {
                        "description": "new info",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serialize.InfobaseInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serialize.InfobaseInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/{app}/infobases": {
            "get": {
                "description": "получение списка информационных баз с сервера 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infobases"
                ],
                "summary": "получение списка информационных баз с сервера 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase user",
                        "name": "infobase-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase password",
                        "name": "infobase-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serialize.InfobaseSummaryInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "description": "создание информационной базы на сервере 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infobases"
                ],
                "summary": "создание информационной базы на сервере 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "create server db",
                        "name": "create-db",
                        "in": "query"
                    },
                    {
                        "description": "new info",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serialize.InfobaseInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serialize.InfobaseInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "serialize.InfobaseInfo": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string",
                    "example": "efa3672f-947a-4d84-bd58-b21997b83561"
                },
                "date_offset": {
                    "type": "integer",
                    "example": 0
                },
                "db_name": {
                    "type": "string",
                    "example": "base"
                },
                "db_pwd": {
                    "type": "string",
                    "example": "password"
                },
                "db_server": {
                    "type": "string",
                    "example": "sql"
                },
                "db_user": {
                    "type": "string",
                    "example": "user"
                },
                "dbms": {
                    "type": "string",
                    "example": "MSSQLServer"
                },
                "denied_from": {
                    "type": "string",
                    "example": "2020-10-01T08:30:00Z"
                },
                "denied_message": {
                    "type": "string",
                    "example": "Выполняется обновление базы"
                },
                "denied_parameter": {
                    "type": "string",
                    "example": "123"
                },
                "denied_to": {
                    "type": "string",
                    "example": "2020-10-01T08:30:00Z"
                },
                "descr": {
                    "type": "string",
                    "example": "Это очень хорошая база"
                },
                "external_session_manager_connection_string": {
                    "type": "string",
                    "example": "http://auth2.com"
                },
                "external_session_manager_required": {
                    "type": "boolean",
                    "example": false
                },
                "license_distribution": {
                    "type": "integer",
                    "example": 0
                },
                "locale": {
                    "type": "string",
                    "example": "ru_RU"
                },
                "name": {
                    "type": "string",
                    "example": "УППБоеваяБаза"
                },
                "permission_code": {
                    "type": "string",
                    "example": "123"
                },
                "reserve_working_processes": {
                    "type": "boolean",
                    "example": false
                },
                "safe_mode_security_profile_name": {
                    "type": "string",
                    "example": "profile1"
                },
                "scheduled_jobs_deny": {
                    "type": "boolean",
                    "example": false
                },
                "security_level": {
                    "type": "integer",
                    "example": 0
                },
                "security_profile_name": {
                    "type": "string",
                    "example": "sec_profile1"
                },
                "sessions_deny": {
                    "type": "boolean",
                    "example": false
                },
                "uuid": {
                    "type": "string",
                    "example": "efa3672f-947a-4d84-bd58-b21997b83561"
                }
            }
        },
        "serialize.InfobaseSummaryInfo": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string"
                },
                "descr": {
                    "type": "string",
                    "example": "УППБоеваяБаза"
                },
                "name": {
                    "type": "string",
                    "example": "УППБоеваяБаза"
                },
                "uuid": {
                    "type": "string",
                    "example": "efa3672f-947a-4d84-bd58-b21997b83561"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "metrics": " Grants read access and control metrics operations",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    }
}